Moved hostname class definition to it's own seperate file
