+++ archey/hostname.py
@@ -1,0 +1,9 @@
+from subprocess import check_output
+
+
+class Hostname:
+    def __init__(self):
+        self.value = check_output(
+            ['uname', '-n'],
+            universal_newlines=True
+        ).rstrip()
